pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strands-telegram-listener"
version = "1.0.0"
description = "Real-time Telegram message processing with AI-powered responses for Strands Agents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Strands Community", email = "community@strandsagents.com"}
]
keywords = [
    "strands", "agents", "telegram", "bot", "listener", "ai", "realtime", 
    "messaging", "automation", "background", "polling", "events"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8", 
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.8"
dependencies = [
    "strands-agents",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.17.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.20.0",
    "responses>=0.23.0",
    "pytest-mock>=3.10.0",
    "pytest-threading>=0.1.0",
]
telegram = [
    "strands-telegram>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/eraykeskinmac/strands-telegram-listener"
Documentation = "https://github.com/eraykeskinmac/strands-telegram-listener#readme"
Repository = "https://github.com/eraykeskinmac/strands-telegram-listener.git"
"Bug Tracker" = "https://github.com/eraykeskinmac/strands-telegram-listener/issues"
Changelog = "https://github.com/eraykeskinmac/strands-telegram-listener/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.yml", "*.yaml"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=strands_telegram_listener",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]